Implement pointing/polarization angle calculation
This is a huge commit that implements the ability to
compute pointing directions and polarization angles.
It affects the way times are kept in Observations:
now if Observation.start_time is a float, it is
*always* interpreted as a time in seconds (previously
its interpretation depended on the value of
Observation.use_mjd). This makes the code less
ambiguous and nicer to write and read.