Implement the "Instrument" class
The "Instrument" class is used to define the orientation of the boresight.
It permits to specify a tilt of the boresight as well as its phase angle
around the spin axis, which means that now the framework is fully able to
support a simulation of the three focal planes (LFT, MFT, HFT).  This comes
at the expense of moving the β parameter (boresight-to-spin-axis angle) out
of the "ScanningStrategy" class and into the "Instrument" class; it solves
an old inconsistency, as the angle β is not part of the «scanning strategy»,
being an hardware-specific parameter.

The dependencies have been updated:

1. The "sphinxcontrib-bibtex" library has been added; it's used to insert
   references to other papers using BibTeX.

2. The "ducc0" dependency now refers to the branch quat_rot_right, as it
   implements a new function needed for the quaternion composition used by
   the code. Once the "quat_rot_right" branch will be merged into "ducc0",
   this dependency will need to be updated accordingly.