Implement Observation.get_ecl2det_quaternions
The new method permits to simulate the observation of pointlike sources in
the sky.  As it depends on a few more quaternion functions, I have moved all
the Numba code dealing with quaternions in the new file
`litebird_sim/quaternions.py`.

Many docstrings have been improved, in particular those associated with
Numba functions.