Rename `Instrument` to `InstrumentInfo`
This is a **breaking** change: the `Instrument` class is renamed to
`InstrumentInfo`, to match the names `DetectorInfo` and
`FreqChannelInfo`. For consistency, the class has been moved from
`instruments.py` to `detectors.py`, which already contained the
definition for classes `DetectorInfo` and `FreqChannelInfo`.

Moreover, the fields used by `InstrumentInfo` to hold angles can no
longer be initialized with values in *degrees*: you must explicitly
pass them as radians. This makes the interface to the class clearer:
the parameters you pass to the constructor are the same you can access
once the object has been constructed.

The mock IMO used in tests has been updated to use the name
`instrument_info` instead of `focal_plane`, to be consistent with the
true IMO. A few new data files have been added to this IMO, to test
the ability to produce data files.

All the examples in the documentation have been updated accordingly.