Add function `detector_list_from_parameters`
This was missing from the previous commit. It implements a
general-purpose way to produce a list of `DetectorInfo` objects out of
a set of definitions in a dictionary, and it is meant to be used with
TOML files.

A new page about detectors, frequency channels, and instruments has
been added to the documentation.