Use only SI units in SpinningScanningStrategy
This breaking change is motivated by issue #65
(https://github.com/litebird/litebird_sim/issues/65). It changes the
name and meaning of the parameters to
`SpinningScanningStrategy.__init__` in the following way:

- `spin_sun_angle_deg` -> `spin_sun_angle_rad` (from degrees to
  radians)

- `spin_rate_rpm` -> `spin_rate_hz` (from RPM to Hertz)

- `precession_period_min` -> `precession_rate_hz` (from a time
  expressed in minutes to a frequency in Hz)

The reason of this is that in this way there is no unit conversion in
the constructor, which leads to a more predictable behavior of the
class. The reason why the internal fields have stayed the same is
because in their current form they ensure the fastest code (i.e., no
conversions during calculations).