Use "dataclasses" (#60)
* Add dependency on dataclasses for py36

* Rename Detector->DetectorInfo and make it a dataclass

Using `@dataclass` vastly simplifies the implementation of `Detector`,
and the new name `DetectorInfo` makes the connection with the IMO
quantity clearer.

* Ensure that detector quaternions are normalized

* Trying to overcome Poetry's bug

Apparently, Poetry 1.0 does not handle dependencies on specific
versions of Python correctly:

https://github.com/python-poetry/poetry/issues/1413

While waiting for Poetry 1.1, this commit tries to fix the problem.

* Add FreqChannelInfo class

* Install a mock IMO in the Travis CI machine

* Add bandcenter and bandwidth to `DetectorInfo`

* Rename `Instrument` to `InstrumentInfo`

This is a **breaking** change: the `Instrument` class is renamed to
`InstrumentInfo`, to match the names `DetectorInfo` and
`FreqChannelInfo`. For consistency, the class has been moved from
`instruments.py` to `detectors.py`, which already contained the
definition for classes `DetectorInfo` and `FreqChannelInfo`.

Moreover, the fields used by `InstrumentInfo` to hold angles can no
longer be initialized with values in *degrees*: you must explicitly
pass them as radians. This makes the interface to the class clearer:
the parameters you pass to the constructor are the same you can access
once the object has been constructed.

The mock IMO used in tests has been updated to use the name
`instrument_info` instead of `focal_plane`, to be consistent with the
true IMO. A few new data files have been added to this IMO, to test
the ability to produce data files.

All the examples in the documentation have been updated accordingly.

* Add function `detector_list_from_parameters`

It implements a general-purpose way to produce a list of `DetectorInfo`
objects out of a set of definitions in a dictionary, and it is meant to be
used with TOML files.

A new page about detectors, frequency channels, and instruments has
been added to the documentation.

* Ensure that pip is upgraded before installing poetry

This might resolve a few problems when installing the
"pillow" dependency under Mac OS X:

https://github.com/python-pillow/Pillow/issues/4242

(the title of the issue is misleading, as the problem
does not affect only Windows user, as it's evident
from the thread.)