Move the CI from Travis to GitHub Actions (#77)
The `.travis.yml` file is replaced with `.github/workflows/tests.yml`. Travis is de-synchronized with the repo and now the tests run on GitHub Actions. Only the minimal changes to achieve this where done. In particular, Windows tests still run on AppVeyor in spite of the fact that they also can be run on GHA