Make sure that distribute_optimally never crashes
There are some peculiar situations (used mostly in unit tests) where
distribute_optimally returns a number of time spans that is smaller
than the number of MPI processes. This happens because of the way
the painter's algorithm works. Suppose that there are 12 observations
of equal length and 9 MPI processes; in this case, the painter's
algorithm comes up with 2 observations for the first 6 processes
and no observations for the last 3. This is an optimal solution
because it minimizes the maximum amount of data per process; any
alternative like 2+2+2+1+1+1+1+1+1 would leave 6 processes IDLE for
50% of the time and still require the same time as 2+2+2+2+2+2+0+0+0.

This commit fixes a failed assert that assumed that all the processes
should be filled with at least one observation.