Refactor years splits and add surveys splits
I arbitrarily decided a maximum amount of years, so that I could refactor and automatize the definitions of the year splits. Of course, this number should be eventually discussed.

Also, I added the survey split. For now the duration of a survey is taken to be 6 months, however, a precise computation might be needed.