Generalized optical chain (#358)
* polarization efficiency and polarization angle introduced

* scan_map restructured

* some checks fixed, only io test broken

* numba implementation fixed

* numba parallelizzation fixed

* some documentation added

* some documentation added

* parallelization removed from scan_map_generic_hwp_for_one_detector

* some more minimal documentation

* pointing matrix updated in mapmakers

* Modify the mock IMo to use pol_angle_rad

All the quaternions are now equal to [0, 0, 0, 1], as the orientation
angle of each of them was really the polarization angle.

* test_simulations fixed

* Improve the tests on the IMo

There was an ambiguity in the expression “mock IMo”, as it could refer
to the PTEP IMo or to the IMo files in the test/ directory. Now the IMo
in the `test/` directory is called `test_imo`, and the directory
containing the YAML file has been renamed accordingly.

* Fix the formatting

* Make write_one_observation work with Mueller matrices too

* Fix the numbers in test_smart_pointings_angles

The new numbers neglects the value of pol_angle_rad. I believe this is
ok, as the three angles θ, φ, ψ should depend only on the geometrical
orientation of the detector and not on the details of its beam.

* Rename `compute_polang_from_detquat` so that `polang` → `orientation`

* update CHANGELOG

* Clarify a docstring

---------

Co-authored-by: Maurizio Tomasi <ziotom78@gmail.com>