Import grasp2alm in LBS (#447)
* Add basic Healpix functionalities

Two notable additions that will be useful to support grasp2alm:

- UNSEEN_PIXEL_VALUE (same as healpy.UNSEEN)
- num_of_alms

* Import grasp2alm’s code from Yusuke’s repository

Instead of keeping classes split into several files, I chose to merge
them into one, to avoid making the litebird_sim/ folder too crowded.

This commit avoids using Healpy, and it employs ducc0.healpix and
ducc0.sht. As a consequence, the interface used to compute the
spherical harmonics has changed, because ducc0.sht does not use
`use_pixel_weights`, `iter`, and so on.

* Add typing for BeamPolar.to_map and grasp2alm()

* Avoid using file names and paths in grasp2alm API

This PR makes sure that you can only load GRASP grid/cut files from
file-like objects instead of paths. In this way, we can more easily use
libinsdb and the IMo. Also, this lets to implement tests without the
need to save files in test/grasp2alm_reference, which could raise errors
if LBS is being tested within a read-only directory. (This sort of
things usually happen within Docker containers.)

* Use `SphericalHarmonics` in grasp2alm

This makes our API diverge once more from grasp2alm, but it makes the
code much easier to use with LBS.

* Remove deprecated parameter

* Update the documentation

* Simplify the code to create a polar map

A few variables were recomputed in a `for` loop, but they were always
the same. Also, there was a private function that was never called in
the code.

* Improve a few docstrings

* Simplify the code that read .cut files and make it faster

* Fix a wrong call to ducc0.healpix

* Update the names of a few members in tests

* Add type int to test

* Add an error message to an assert

* Fix tests and use a more representative .cut file

* Make TestBeamCut test pass

* Make all tests pass

* Restructure BeamGrid and BeamPolar to support IGRID≠7

This is a WIP, as the internal representation for BeamPolarStokes must
switch to a list of (θ, φ) values instead of two linear arrays for θ and
for φ.

* Make the code accept azimuth/elevation grid files

The commit contains an extensive test case developed using Mathematica.

* Modularize tests in grasp2alm and add comments

* Add two functions to get the solid angle and side size of Healpix pixels

The two functions are `nside_to_pixel_solid_angle_sterad` and
`nside_to_resolution_rad`.

* Export the two new Healpix functions

* Use a binner to compute the Stokes map of a beam

The old approach used an interpolator, but for non-uniform grids like
the azimuth/elevation IGRID=5 schema used by GRASP the code did not
work. This new approach uses a plain binning algorithm to convert a
`BeamStokesPolar` into a `BeamMap` in `grasp2alm`.

* Improve compatibility with ducc0.sht

* Fix math in docstring for BeamStokesPolar

* Add some documentation to the user's manual

* [skipci] Update CHANGELOG